// src/lib/models/RecommendationResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecommendationResponse = {
    id: number;
    titre: string;
    description: string;
    date_creation: string;
    parcelle_id: number;
    created_at: string;
    updated_at: string;
};